fix(kotlin): resolve remaining compileDebugKotlin errors across 6 files
These surfaced once the earlier resource-merge and brace-balance blockers
were cleared, letting the compiler reach real type errors:

- FastingViewModel: personalRecord StateFlow declared Int but
  FastCompletion.achievedHours is Double — fixed the declared type
  instead of truncating the data.
- HydrationScreen / FastingScreen: missing androidx.compose.ui.draw.clip
  import used by the 7-day chart bars.
- TrackerScreen: duplicate RoundedCornerShape import made it ambiguous.
- MetabolicHealthScoreCard: met.npRq doesn't exist on MetabolicResult —
  npRq lives on the nested met.sub (SubstrateResult).
- MealPlanViewModel: MealTemplate has no .components; use the existing
  .totalKcal computed property instead.
- ProfileScreen: Goal enum values are LOSE/MAINTAIN/GAIN, not
  LOSE_WEIGHT/GAIN_MUSCLE.

+++ scan-eat-android/app/src/main/java/fr/scanneat/presentation/fasting/FastingViewModel.kt
@@ -26,8 +26,8 @@ class FastingViewModel @Inject constructor(
     val language: StateFlow<String> = prefs.language
         .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), "fr")
     /** Longest achieved fast in hours — used to surface a personal-record alert. */
-    val personalRecord: StateFlow<Int> = history.map { list -> list.maxOfOrNull { it.achievedHours } ?: 0 }
-        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), 0)
+    val personalRecord: StateFlow<Double> = history.map { list -> list.maxOfOrNull { it.achievedHours } ?: 0.0 }
+        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), 0.0)
 
     // Tick every second so the UI re-draws the elapsed counter
     val tick: StateFlow<Long> = flow {
